src/test/resources/docker/numcites/requirements.txt
bibtexparser
numpy
scholarly
python-Levenshtein